simplify compiler syntax naming
