+++ src/compiler/codegen/emitter.rs
@@ -12,7 +12,7 @@ use crate::compiler::model::{
     spawn_target_state,
 };
 use crate::compiler::syntax::lexer::{RESERVED_GENERATED_PREFIX, RESERVED_GENERATED_TYPE_PREFIX, TokenKind, lex};
-use crate::compiler::syntax::words::word;
+use crate::compiler::syntax::word;
 use crate::compiler::syntax::*;
 use crate::error::{ArgentError, Result};
 use crate::naming::{is_identifier, to_snake};
