+++ src/compiler/syntax/word.rs
@@ -1,0 +1,39 @@
+//! Canonical spellings for language words interpreted directly by Argent.
+
+pub const ACTOR: &str = "actor";
+pub const ACTOR_TYPE: &str = "actor_type";
+pub const APP: &str = "app";
+pub const AS: &str = "as";
+pub const BECOME: &str = "become";
+pub const BY: &str = "by";
+pub const CO_SPENT: &str = "co_spent";
+pub const CONSUMES: &str = "consumes";
+pub const CONST: &str = "const";
+pub const COVENANT_ID: &str = "cov_id";
+pub const LEGACY_COVENANT_ID: &str = "covid";
+pub const DELEGATE: &str = "delegate";
+pub const ELSE: &str = "else";
+pub const EMITS: &str = "emits";
+pub const ENTRY: &str = "entry";
+pub const ENUM: &str = "enum";
+pub const EXPANDS: &str = "expands";
+pub const FN: &str = "fn";
+pub const FOR: &str = "for";
+pub const FROM: &str = "from";
+pub const IF: &str = "if";
+pub const IMPORT: &str = "import";
+pub const INPUTS: &str = "inputs";
+pub const LEADER: &str = "leader";
+pub const NONE: &str = "none";
+pub const OBSERVES: &str = "observes";
+pub const ONE: &str = "one";
+pub const OUTPUTS: &str = "outputs";
+pub const OWNS: &str = "owns";
+pub const REQUIRE: &str = "require";
+pub const RESERVED_SELF_MEMBERS: [&str; 5] = [STATE, VALUE, COVENANT_ID, "type", "ref"];
+pub const SELF: &str = "self";
+pub const SPAWNS: &str = "spawns";
+pub const STATE: &str = "state";
+pub const UNRESTRICTED: &str = "unrestricted";
+pub const VALUE: &str = "value";
+pub const VIRTUAL: &str = "virtual";
